+++ examples/minimum/main.tf
@@ -12,11 +12,10 @@ module "minimal" {
   availability_zones = data.aws_availability_zones.available.names
 
   # Launch template
-  launch_template_name        = local.name
   launch_template_description = "minimal launch template example"
   create_launch_template      = true
   image_id                    = data.aws_ami.amazon_linux.id
-  instance_type               = "t2.micro"
+  instance_type               = "t3.nano"
 
   metadata_options = {
     http_endpoint = "enabled"